Added comparing of windows to sessions.
The `checkIfExistingSession` method will compare the passed window object with the sessions saved in the 'Seshy' bookmark folder to see if the window represents an existing session.
It does this by comparing each tab of the window with each bookmark of the bookmark folder.
If all a windows tabs are equal to the bookmarks in a folder, the window is considered an existing session.