Adding build process. WIP.
Cannot find an easy way to have a test process for Chrome extensions because I do not know of a way to reference JS files to test from parent or sibling directories in relation to the mainfest file.
Could have all the test and implementation files lumped in together but then there is no way to publish without test files included which is bad.
Also event listeners need to be disabled for the tests otherwise methods under test are firing during setup of test when windows are created.
There's no easy way to disable the event listeners without passing a flag to say 'testing', but do not want implementation to know about test.

A build process allows a 'test extension' to be assembled just for the purposes of testing and will provide additional flexibility for publishing and other tasks.