Subsequent saves overwrite session folder.
Before multiple saves caused new bookmarks to be written in the same folder and then new session folders to be created which was unintended.
Now a mapping of window ids to session folder ids is stored in Chrome's local storage when a session is first saved so that subsequent session saves overwrite the same folder.
This is the mechanism that will be used to updated the folder continuously as the session evolves.