Session manager refactors.
Removed fake 'Recently Closed' section, it is to be implemented properly in a future release.
Added docstrings, but am not yet using JSDocs.
Removed `initialise()` call in the session manager JS as it is not certain that it is necessary. It will be called everytime the `init.js` background page is loaded up, the question is what exactly triggers this to be loaded? Does any call to a method in a background page wake it up?