Renamed 'Saved Sessions' to 'Shelved Sessions'.
Having a 'Saved Sessions' list implied that saved sessions appeared exclusively in that list, but saved sessions can also appear in the 'Currently Open Sessions' list. The word 'Shelved' is less colloquial but much accurate. The sections will be explained in the Welcome Page so a slightly less universally understood word should not cause much harm.