Added custom Jasmine reporter.
The Jasmine reporter receives the spec results as an object and overwrites the pages HTML to display information about the spec results in a simpler format. The intention is that the page can be downloaded and then parsed with a simple script to determine whether the Travis build passed or failed. Also updated Jasmine version.