Attempting new method for getting test results.
In the previous attempt a custom Jasmine reporter was used which could reliably get the test results without nasty hacks like DOM inspection, but then had no reliable way of recording the results somewhere.
It is necessary to record the results somewhere so that another script can read the results and return an appropriate exit code to pass or fail Travis. This is fundamentally because the reporter is operating in the browser and so has limited access to the host system (the Travis instance).

The new method is to use Chrome Driver to drive the browser which will run the tests using Jasmine runner just as it did before. The difference is that the controlling scope is Chrome Driver and not the browser so greater access to the host system is available and so there should be plenty of options for passing / failing Travis.