Tests now initiated by ChromeDriver.
With ChromeDriver we are in NodeJS land so it is easy to exit the script with a zero exit code if the tests pass or a non-zero exit code if the tests fail.

One niggle is that the browser window is left open when the tests fail. Will address that in a future commit.