Created an init file for the session manager.
Having a separate init file make it easier to test the functions without having to have the correct DOM in place.
For example the session manager `setup()` function expects certain elements to be in the DOM and therefore fails when run as part of the spec runner HTML that has different elements entirely.