Made test setup function redundant.
Simplified test setup by making a commonly called function `clearLocalStorageAndInitialise` redundant.
Clearing of local storage should be the responsibility of the test who created data there in the first place, so moved this to `cleanUp` function.
Initialisation was required to recreate the `Seshy` bookmark folder that is the root for all other session folders as that is removed during `cleanUp` after every test.
Changed this so that the `Seshy` folder is never removed, only its child folders.