Experimented with using Jasmine's context.
In Jasmine functions `this` is bound to an instance of Jasmine's UserContext class. This same instance is bound in call Jasmine functions within a single spec so it is a convenient place to put test vairables.
Additional benefits of UserContext are that a new instance is used in each spec so cleanup of the variable namespace is done for you, preventing leaking.

Using this also presented an opportunity to explore the differences between function expressions and function declarations in practice.
Some comments were added around this issue for learning purposes.