Various refactoring.
Added some tests.
Added a hacky implementation for saving that does not work properly. This is going to be done properly in upcoming commits.
Renamed some methods to distinguish between the terms 'saved' and 'shelved'. Any session that is 'saved' will be automatically 'shelved' when closed. Currently 'shelved' sessions are those that are saved but not open. Currently open 'saved' sessions are 'unshelved'.