Removed redundant Grunt tasks.
Previously before running tests all files would be copied to an output directory and then Chrome's CLI would be used to start an instance with an extension loaded from that directory.
That was necessary to use the test manifest file selectively with some of the implementation files and all the test files.
The Grunt CRX plugin renders this unnecessary as it allows you to specify source files to be copied into the CRX package that it creates. The structure is flattened which is not ideal but it is still prederable due to the steps eliminated from the process.