List element focused instead of input.
When navigating with the arrow keys the session name input would be focused. This was bad because it means that all keyboard input is captured by the text field therefore disabling any keyboard shortcuts.
Focusing the session card instead and allowing the users to decide when they want to rename the session with the `r` is a good compromise.
Had to add `tabindex=0` as an attribute on the session list elements in order to make them focusable.