Test for session naming and refactoring.
Added test for the name of shelved sessions. Still need to add one for currently open saved sessions.
Tests run much quicker now because the `cleanUp` method no longer relies on a `setTimeout`, instead it callsback as soon as the `cleanUp` is done.
Extracted the `asyncLoop` method to a `util.js` module. This file will hold methods that are used in both the implementation and tests, but will also be themselves tested. This avoids duplicating obviously useful methods that will naturally occur in both the implementation and tests due to encapsulation.