Fix sessions disappearing
When Chrome was quit the open saved session mapping in local storage was not being cleaned up which meant that next time Chrome was opened the extension thought it was still unshelved but with a window ID that is no longer being used. Due to the logic in the code this resulted in the session not showing up at all in the UI.

We fix this by running an existing function to remove stale mappings on startup.

Fixes #102.